moved DidntFollowInstructionsAppSetup to own class
