+++ src/Shimmer.Client/UpdateManager.cs
@@ -1,6 +1,5 @@
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Diagnostics.Contracts;
 using System.IO.Abstractions;
 using System.Linq;
@@ -698,24 +697,6 @@ namespace Shimmer.Client
                 .SelectMany(x => Utility.DeleteDirectory(x.FullName, RxApp.TaskpoolScheduler))
                     .LoggedCatch<Unit, UpdateManager, UnauthorizedAccessException>(this, _ => Observable.Return(Unit.Default))
                 .Aggregate(Unit.Default, (acc, x) => acc);
-        }
-    }
-
-    public class DidntFollowInstructionsAppSetup : AppSetup
-    {
-        readonly string shortCutName;
-        public override string ShortcutName {
-            get { return shortCutName; }
-        }
-
-        readonly string target;
-        public override string Target { get { return target; } }
-
-        public DidntFollowInstructionsAppSetup(string exeFile)
-        {
-            var fvi = FileVersionInfo.GetVersionInfo(exeFile);
-            shortCutName = fvi.ProductName ?? fvi.FileDescription ?? fvi.FileName.Replace(".exe", "");
-            target = exeFile;
         }
     }
 }
